Adjusting docker compse configurations
Now, the databases are perpetually stored in external space out of the docker. Thus, the database files are kept permanently.

Additionally, made some conditional stuff to detect in-docker situation automatically to dynamically set the redis name(conventionally, "localhost" is the host name of the Windows local REDIS server whereas my docker-compose.yml configured it as "db-redis" for more readability.)