Profile picture mismanagement bug fix
The image was stored in wrong way(a stupid code of user_id = 1 hardcode etc.). Refer to the issue #1 . Also, I managed the pictures with user's UUID, but it's counterintuitive and unnecessary since the username's uniqueness is preserved by the registration step. Now we save the images with the username. (i.e. "122a9c0-.....32c.png" --> "jaclyn000.png")